+++ Plugin/RLTrainingTimer.vcxproj
@@ -95,12 +95,12 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="adapter\RocketLeagueEventAdapter.cpp" />
-    <ClCompile Include="configuration\ConfigurationUi.cpp" />
-    <ClCompile Include="configuration\TrainingProgramConfigurationControl.cpp" />
-    <ClCompile Include="configuration\TrainingProgramConfigurationUi.cpp" />
-    <ClCompile Include="configuration\TrainingProgramListConfigurationControl.cpp" />
-    <ClCompile Include="configuration\TrainingProgramListConfigurationUi.cpp" />
-    <ClCompile Include="configuration\TrainingProgramPersistenceService.cpp" />
+    <ClCompile Include="configuration\control\TrainingProgramConfigurationControl.cpp" />
+    <ClCompile Include="configuration\control\TrainingProgramListConfigurationControl.cpp" />
+    <ClCompile Include="configuration\control\TrainingProgramPersistenceService.cpp" />
+    <ClCompile Include="configuration\ui\ConfigurationUi.cpp" />
+    <ClCompile Include="configuration\ui\TrainingProgramConfigurationUi.cpp" />
+    <ClCompile Include="configuration\ui\TrainingProgramListConfigurationUi.cpp" />
     <ClCompile Include="core\kernel\InvalidStateException.cpp" />
     <ClCompile Include="core\training\application\TrainingApplicationService.cpp" />
     <ClCompile Include="core\training\application\TrainingProgramDisplayReadModel.cpp" />
@@ -133,15 +133,16 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="adapter\RocketLeagueEventAdapter.h" />
-    <ClInclude Include="configuration\ConfigurationUi.h" />
-    <ClInclude Include="configuration\event_serializing.hpp" />
-    <ClInclude Include="configuration\json_serializer.hpp" />
-    <ClInclude Include="configuration\TrainingProgramConfigurationControl.h" />
-    <ClInclude Include="configuration\TrainingProgramConfigurationUi.h" />
-    <ClInclude Include="configuration\TrainingProgramEntry.h" />
-    <ClInclude Include="configuration\TrainingProgramListConfigurationControl.h" />
-    <ClInclude Include="configuration\TrainingProgramListConfigurationUi.h" />
-    <ClInclude Include="configuration\TrainingProgramPersistenceService.h" />
+    <ClInclude Include="configuration\control\event_serializing.hpp" />
+    <ClInclude Include="configuration\control\json_serializer.hpp" />
+    <ClInclude Include="configuration\control\TrainingProgramConfigurationControl.h" />
+    <ClInclude Include="configuration\control\TrainingProgramListConfigurationControl.h" />
+    <ClInclude Include="configuration\control\TrainingProgramPersistenceService.h" />
+    <ClInclude Include="configuration\data\TrainingProgramData.h" />
+    <ClInclude Include="configuration\data\TrainingProgramEntry.h" />
+    <ClInclude Include="configuration\ui\ConfigurationUi.h" />
+    <ClInclude Include="configuration\ui\TrainingProgramConfigurationUi.h" />
+    <ClInclude Include="configuration\ui\TrainingProgramListConfigurationUi.h" />
     <ClInclude Include="core\kernel\InvalidStateException.h" />
     <ClInclude Include="core\training\application\TrainingApplicationService.h" />
     <ClInclude Include="core\training\application\TrainingProgramDisplayReadModel.h" />
@@ -183,7 +184,6 @@
     <ClInclude Include="pch.h" />
     <ClInclude Include="qol\vectorext.h" />
     <ClInclude Include="RLTrainingTimer.h" />
-    <ClInclude Include="configuration\TrainingProgramData.h" />
     <ClInclude Include="ui\training\TrainingProgramDisplay.h" />
     <ClInclude Include="ui\training\TrainingProgramFlowControlUi.h" />
     <ClInclude Include="version.h" />
