+++ Plugin/configuration/control/TrainingProgramConfigurationControl.cpp
