+++ Plugin/configuration/control/TrainingProgramListConfigurationControl.cpp
