+++ Plugin/configuration/control/TrainingProgramPersistenceService.cpp
