+++ Plugin/configuration/control/json_serializer.hpp
