+++ Plugin/configuration/data/TrainingProgramEntry.h
