+++ Plugin/configuration/ui/TrainingProgramConfigurationUi.cpp
