+++ Plugin/configuration/ui/TrainingProgramListConfigurationUi.cpp
