+++ Plugin/version.h
@@ -2,7 +2,7 @@
 #define VERSION_MAJOR 0
 #define VERSION_MINOR 1
 #define VERSION_PATCH 0
-#define VERSION_BUILD 477
+#define VERSION_BUILD 482
 
 #define stringify(a) stringify_(a)
 #define stringify_(a) #a
